Allow you to set a sensor Address like you can for vXL53..
However I only added the setAddress, and also added a getAddress to retrieve it later if desired.

I did not add it to begin as that would make the begin not be compatible with existing code.

Added a keywords.txt file

Added a three sensor test like the dual for the other one...
Nothing special but appears to work.