Asynchronous read ranges
Experimenting with adding appis that allow you to instead of do a readRange, which starts it up and waits for it to complete, instead add a set of APIs that allow me
to startup a range(Adafruit_VL6180X::startRange), check to see if it is done (isRangeComplete), wait for it to complete (waitRangeComplete)

and when completed ou call readRangeResult, which returns the value and dlears out the system interrupt.

I also updated the triple sensor test program, with real simple serial input to change modes
d - back to the earlier output with more detailed output.
t - timed output for reading each of the three servos one at a time
a - async timed start all three wait for all three to complete and grab values.

Keywords.txt also updated to show the new methods.

Not sure if you wish to review these names before merging.  Also may also look at maybe looking at some other functionality...   Which could then be added to this PULL or could be done on feature by feature...

Hopefully I did not do any doxygen... or clang issues. but will watch to see as it is verified.