Add continuous range support also updated example sketch
Not sure if I took this far enough, as I only added continuous support for ranging, not the LUX reads, which also could be made to either or both...

But I also revamed the triple sketch, which has quick and dirty one character mode settings to have one mode for doing continuous.

Also in this version of the sketch is a timed version using GPIO pins for testing completion versus the polling method.  So far I only tested the GPIO way with the continuous mode, but added support for the polling way as well.

Probably getting much done with this round of playing around, especially since I am not sure if I will actully use these sensors, but has been fun.