Fix the ALS intergration period in init sequence
According to AN4545, the integration period for ALS measurements (16-bit register 0x40) is set to 100 ms, which corresponds to register value 0x0063. Since 16-bit registers are big-endian, the low byte must be written after 0x41 as specified in the init sequence in AN4545. The high byte at address 0x40 is already 0 after a reset.